+++ client/spammer.go
@@ -12,11 +12,11 @@ const (
 )
 
 // ToggleSpammer toggles the node internal spammer.
-func (api *GoShimmerAPI) ToggleSpammer(enable bool) (*webapi_spammer.Response, error) {
+func (api *GoShimmerAPI) ToggleSpammer(enable bool, mpm int) (*webapi_spammer.Response, error) {
 	res := &webapi_spammer.Response{}
 	if err := api.do(http.MethodGet, func() string {
 		if enable {
-			return fmt.Sprintf("%s?cmd=start", routeSpammer)
+			return fmt.Sprintf("%s?cmd=start&mpm=%d", routeSpammer, mpm)
 		}
 		return fmt.Sprintf("%s?cmd=stop", routeSpammer)
 	}(), nil, res); err != nil {