+++ client/value.go
@@ -62,21 +62,23 @@ func (api *GoShimmerAPI) SendTransaction(txnBytes []byte) (string, error) {
 	return res.TransactionID, nil
 }
 
-// SendTransactionByJSON sends the transaction(JSON) to the Value Tangle and returns transaction ID.
-//func (api *GoShimmerAPI) SendTransactionByJSON(txn webapi_value.SendTransactionByJSONRequest) (string, error) {
-//	res := &webapi_value.SendTransactionByJSONResponse{}
-//	if err := api.do(http.MethodPost, routeSendTxnByJSON,
-//		&webapi_value.SendTransactionByJSONRequest{
-//			Inputs:     txn.Inputs,
-//			Outputs:    txn.Outputs,
-//			Data:       txn.Data,
-//			Signatures: txn.Signatures,
-//		}, res); err != nil {
-//		return "", err
-//	}
-//
-//	return res.TransactionID, nil
-//}
+// SendTransactionByJSON sends the transaction(JSON) to the Value Tangle and returns transaction ID and message ID.
+func (api *GoShimmerAPI) SendTransactionByJSON(txn webapi_value.SendTransactionByJSONRequest) (string, error) {
+	res := &webapi_value.SendTransactionByJSONResponse{}
+	if err := api.do(http.MethodPost, routeSendTxnByJSON,
+		&webapi_value.SendTransactionByJSONRequest{
+			Inputs:        txn.Inputs,
+			Outputs:       txn.Outputs,
+			AManaPledgeID: txn.AManaPledgeID,
+			CManaPledgeID: txn.CManaPledgeID,
+			Signatures:    txn.Signatures,
+			Payload:       txn.Payload,
+		}, res); err != nil {
+		return "", err
+	}
+
+	return res.TransactionID, nil
+}
 
 // GetAllowedManaPledgeNodeIDs returns the list of allowed mana pledge IDs.
 func (api *GoShimmerAPI) GetAllowedManaPledgeNodeIDs() (*webapi_value.AllowedManaPledgeResponse, error) {