+++ packages/tangle/utils.go
@@ -32,7 +32,7 @@ func NewUtils(tangle *Tangle) (utils *Utils) {
 // can be used to control the behavior of the walk similar to how a "Context" is used in some parts of the stdlib.
 func (u *Utils) WalkMessageID(callback func(messageID MessageID, walker *walker.Walker), entryPoints MessageIDs, revisitElements ...bool) {
 	if len(entryPoints) == 0 {
-		panic("you need to provide at least one entry point")
+		return
 	}
 
 	messageIDWalker := walker.New(revisitElements...)
@@ -89,75 +89,51 @@ func (u *Utils) WalkMessageAndMetadata(callback func(message *Message, messageMe
 
 // region structural checks ////////////////////////////////////////////////////////////////////////////////////////////
 
-// AllTransactionsApprovedByMessage checks if all Transactions were attached by at least one Message that was directly
+// AllTransactionsApprovedByMessages checks if all Transactions were attached by at least one Message that was directly
 // or indirectly approved by the given Message.
-func (u *Utils) AllTransactionsApprovedByMessage(transactionIDs ledgerstate.TransactionIDs, messageID MessageID) (approved bool) {
-	for transactionID := range transactionIDs {
-		if !u.TransactionApprovedByMessage(transactionID, messageID) {
-			return false
-		}
-	}
-
-	return true
-}
-
-// AllTransactionsContainedOrApprovedByMessages checks if all Transactions were attached by at least one Message that was directly
-// or indirectly approved by the given Messages.
-func (u *Utils) AllTransactionsContainedOrApprovedByMessages(transactionIDs ledgerstate.TransactionIDs, messageIDs MessageIDs) (approved bool) {
-	// keep track of already approved transactions
-	approvedTransactions := make(map[ledgerstate.TransactionID]bool)
-	for transactionID := range transactionIDs {
-		approvedTransactions[transactionID] = false
-	}
-
-	// check if transactions are contained in messages
-	for _, messageID := range messageIDs {
-		for transactionID := range transactionIDs {
-			// no need to check if it's already contained in another message
-			if approvedTransactions[transactionID] {
-				continue
-			}
-
-			approvedTransactions[transactionID] = u.tangle.Storage.IsTransactionAttachedByMessage(transactionID, messageID)
-		}
-	}
+func (u *Utils) AllTransactionsApprovedByMessages(transactionIDs ledgerstate.TransactionIDs, messageIDs ...MessageID) (approved bool) {
+	transactionIDs = transactionIDs.Clone()
 
 	for _, messageID := range messageIDs {
 		for transactionID := range transactionIDs {
-			// no need to check if it's already approved by another message
-			if approvedTransactions[transactionID] {
-				continue
+			if u.TransactionApprovedByMessage(transactionID, messageID) {
+				delete(transactionIDs, transactionID)
 			}
-
-			//fmt.Println("Checking with message", messageID)
-			approvedTransactions[transactionID] = u.TransactionApprovedByMessage(transactionID, messageID)
 		}
 	}
 
-	for transactionID := range approvedTransactions {
-		if !approvedTransactions[transactionID] {
-			return false
-		}
-	}
-	return true
+	return len(transactionIDs) == 0
 }
 
 // TransactionApprovedByMessage checks if the Transaction was attached by at least one Message that was directly or
 // indirectly approved by the given Message.
 func (u *Utils) TransactionApprovedByMessage(transactionID ledgerstate.TransactionID, messageID MessageID) (approved bool) {
 	for _, attachmentMessageID := range u.tangle.Storage.AttachmentMessageIDs(transactionID) {
-		if u.tangle.Utils.MessageApprovedByStrongParents(attachmentMessageID, messageID) {
+		if attachmentMessageID == messageID {
 			return true
+		}
+
+		u.tangle.Storage.Message(messageID).Consume(func(message *Message) {
+			for _, parentID := range message.StrongParents() {
+				if u.MessageApprovedBy(attachmentMessageID, parentID) {
+					approved = true
+					return
+				}
+			}
+		})
+
+		if approved {
+			return
 		}
 	}
 
-	return false
+	return
 }
 
 // MessageApprovedBy checks if the Message given by approvedMessageID is directly or indirectly approved by the
 // Message given by approvingMessageID.
 func (u *Utils) MessageApprovedBy(approvedMessageID MessageID, approvingMessageID MessageID) (approved bool) {
-	if u.MessageStronglyApprovedBy(approvedMessageID, approvingMessageID) {
+	if u.messageStronglyApprovedBy(approvedMessageID, approvingMessageID) {
 		return true
 	}
 
@@ -169,7 +145,7 @@ func (u *Utils) MessageApprovedBy(approvedMessageID MessageID, approvingMessageI
 			continue
 		}
 
-		if u.MessageStronglyApprovedBy(weakApprover.ApproverMessageID(), approvingMessageID) {
+		if u.messageStronglyApprovedBy(weakApprover.ApproverMessageID(), approvingMessageID) {
 			return true
 		}
 	}
@@ -177,22 +153,20 @@ func (u *Utils) MessageApprovedBy(approvedMessageID MessageID, approvingMessageI
 	return false
 }
 
-// MessageApprovedByStrongParents checks if the Message given by approvedMessageID is directly or indirectly approved by its strong parents.
-func (u *Utils) MessageApprovedByStrongParents(approvedMessageID MessageID, approvingMessageID MessageID) (approved bool) {
-	u.tangle.Storage.Message(approvingMessageID).Consume(func(message *Message) {
-		for _, parentID := range message.StrongParents() {
-			if u.MessageApprovedBy(approvedMessageID, parentID) {
-				approved = true
-				return
-			}
-		}
+// ApprovingMessageIDs returns the MessageIDs that approve a given Message. It accepts an optional ApproverType to
+// filter the Approvers.
+func (u *Utils) ApprovingMessageIDs(messageID MessageID, optionalApproverType ...ApproverType) (approvingMessageIDs MessageIDs) {
+	approvingMessageIDs = make(MessageIDs, 0)
+	u.tangle.Storage.Approvers(messageID, optionalApproverType...).Consume(func(approver *Approver) {
+		approvingMessageIDs = append(approvingMessageIDs, approver.ApproverMessageID())
 	})
+
 	return
 }
 
-// MessageStronglyApprovedBy checks if the Message given by approvedMessageID is directly or indirectly approved by the
+// messageStronglyApprovedBy checks if the Message given by approvedMessageID is directly or indirectly approved by the
 // Message given by approvingMessageID (ignoring weak parents as a potential last reference).
-func (u *Utils) MessageStronglyApprovedBy(approvedMessageID MessageID, approvingMessageID MessageID) (stronglyApproved bool) {
+func (u *Utils) messageStronglyApprovedBy(approvedMessageID MessageID, approvingMessageID MessageID) (stronglyApproved bool) {
 	if approvedMessageID == approvingMessageID || approvedMessageID == EmptyMessageID {
 		return true
 	}
@@ -235,17 +209,6 @@ func (u *Utils) MessageStronglyApprovedBy(approvedMessageID MessageID, approving
 			})
 		}, u.ApprovingMessageIDs(approvedMessageID, StrongApprover))
 	}
-
-	return
-}
-
-// ApprovingMessageIDs returns the MessageIDs that approve a given Message. It accepts an optional ApproverType to
-// filter the Approvers.
-func (u *Utils) ApprovingMessageIDs(messageID MessageID, optionalApproverType ...ApproverType) (approvingMessageIDs MessageIDs) {
-	approvingMessageIDs = make(MessageIDs, 0)
-	u.tangle.Storage.Approvers(messageID, optionalApproverType...).Consume(func(approver *Approver) {
-		approvingMessageIDs = append(approvingMessageIDs, approver.ApproverMessageID())
-	})
 
 	return
 }