+++ plugins/dashboard/frontend/src/app/components/TransactionPayload.tsx
@@ -8,7 +8,6 @@ import {inject, observer} from "mobx-react";
 import {ExplorerStore} from "app/stores/ExplorerStore";
 import {IconContext} from "react-icons"
 import {FaChevronCircleRight} from "react-icons/fa"
-import * as dateformat from 'dateformat';
 
 interface Props {
     explorerStore?: ExplorerStore;
@@ -51,7 +50,7 @@ export class TransactionPayload extends React.Component<Props, any> {
                 <Row className={"mb-3"}>
                     <Col>
                         <ListGroup>
-                            <ListGroup.Item>Timestamp: {dateformat(new Date(payload.tx_essence.timestamp*1000).toLocaleString(), "dd.mm.yyyy HH:MM:s")} </ListGroup.Item>
+                            <ListGroup.Item>Timestamp: {new Date(payload.tx_essence.timestamp*1000).toLocaleString()} </ListGroup.Item>
                             <ListGroup.Item>Access Pledge ID: {payload.tx_essence.access_pledge_id}</ListGroup.Item>
                             <ListGroup.Item>Consensus Pledge ID: {payload.tx_essence.access_pledge_id}</ListGroup.Item>
                         </ListGroup>