+++ plugins/metrics/plugin.go
@@ -126,6 +126,7 @@ func registerLocalMetrics() {
 			// MessageStored is triggered in storeMessageWorker that saves the msg to database
 			messageTotalCountDB.Inc()
 		})
+		increasePerComponentCounter(Store)
 	}))
 
 	messagelayer.Tangle().Storage.Events.MessageRemoved.Attach(events.NewClosure(func(messageId tangle.MessageID) {
@@ -135,6 +136,7 @@ func registerLocalMetrics() {
 
 	// messages can only become solid once, then they stay like that, hence no .Dec() part
 	messagelayer.Tangle().Solidifier.Events.MessageSolid.Attach(events.NewClosure(func(messageID tangle.MessageID) {
+		increasePerComponentCounter(Solidifier)
 		solidTimeMutex.Lock()
 		defer solidTimeMutex.Unlock()
 
@@ -155,6 +157,14 @@ func registerLocalMetrics() {
 	// fired when a missing message was received and removed from missing message storage
 	messagelayer.Tangle().Storage.Events.MissingMessageStored.Attach(events.NewClosure(func(tangle.MessageID) {
 		missingMessageCountDB.Dec()
+	}))
+
+	messagelayer.Tangle().Scheduler.Events.MessageScheduled.Attach(events.NewClosure(func(messageID tangle.MessageID) {
+		increasePerComponentCounter(Scheduler)
+	}))
+
+	messagelayer.Tangle().Booker.Events.MessageBooked.Attach(events.NewClosure(func(message tangle.MessageID) {
+		increasePerComponentCounter(Booker)
 	}))
 
 	// // Value payload attached