+++ plugins/prometheus/tangle.go
@@ -6,14 +6,15 @@ import (
 )
 
 var (
-	messageTips           prometheus.Gauge
-	messagePerTypeCount   *prometheus.GaugeVec
-	messageTotalCount     prometheus.Gauge
-	messageTotalCountDB   prometheus.Gauge
-	messageSolidCountDB   prometheus.Gauge
-	avgSolidificationTime prometheus.Gauge
-	messageMissingCountDB prometheus.Gauge
-	messageRequestCount   prometheus.Gauge
+	messageTips              prometheus.Gauge
+	messagePerTypeCount      *prometheus.GaugeVec
+	messagePerComponentCount *prometheus.GaugeVec
+	messageTotalCount        prometheus.Gauge
+	messageTotalCountDB      prometheus.Gauge
+	messageSolidCountDB      prometheus.Gauge
+	avgSolidificationTime    prometheus.Gauge
+	messageMissingCountDB    prometheus.Gauge
+	messageRequestCount      prometheus.Gauge
 
 	transactionCounter prometheus.Gauge
 	valueTips          prometheus.Gauge
@@ -31,6 +32,14 @@ func registerTangleMetrics() {
 			Help: "number of messages per payload type seen since the start of the node",
 		}, []string{
 			"message_type",
+		})
+
+	messagePerComponentCount = prometheus.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Name: "tangle_messages_per_component_count",
+			Help: "number of messages per component seen since the start of the node",
+		}, []string{
+			"component",
 		})
 
 	messageTotalCount = prometheus.NewGauge(prometheus.GaugeOpts{
@@ -70,6 +79,7 @@ func registerTangleMetrics() {
 
 	registry.MustRegister(messageTips)
 	registry.MustRegister(messagePerTypeCount)
+	registry.MustRegister(messagePerComponentCount)
 	registry.MustRegister(messageTotalCount)
 	registry.MustRegister(messageTotalCountDB)
 	registry.MustRegister(messageSolidCountDB)
@@ -86,6 +96,10 @@ func collectTangleMetrics() {
 	msgCountPerPayload := metrics.MessageCountSinceStartPerPayload()
 	for payloadType, count := range msgCountPerPayload {
 		messagePerTypeCount.WithLabelValues(payloadType.String()).Set(float64(count))
+	}
+	msgCountPerComponent := metrics.MessageCountSinceStartPerComponent()
+	for component, count := range msgCountPerComponent {
+		messagePerComponentCount.WithLabelValues(component.String()).Set(float64(count))
 	}
 	messageTotalCount.Set(float64(metrics.MessageTotalCountSinceStart()))
 	messageTotalCountDB.Set(float64(metrics.MessageTotalCountDB()))