+++ plugins/webapi/value/sendtransaction.go
@@ -5,7 +5,9 @@ import (
 	"fmt"
 	"net/http"
 	"sync"
+	"time"
 
+	"github.com/iotaledger/goshimmer/packages/clock"
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/packages/mana"
 	"github.com/iotaledger/goshimmer/packages/tangle"
@@ -53,6 +55,33 @@ func sendTransactionHandler(c echo.Context) error {
 				Error: fmt.Errorf("not allowed to pledge consensus mana to %s: %w", tx.Essence().ConsensusPledgeID().String(), ErrNotAllowedToPledgeManaToNode).Error(),
 			})
 		}
+	}
+
+	// check balances validity
+	consumedOutputs := make(ledgerstate.Outputs, len(tx.Essence().Inputs()))
+	for i, consumedOutputID := range tx.Essence().Inputs() {
+		referencedOutputID := consumedOutputID.(*ledgerstate.UTXOInput).ReferencedOutputID()
+		messagelayer.Tangle().LedgerState.Output(referencedOutputID).Consume(func(output ledgerstate.Output) {
+			consumedOutputs[i] = output
+		})
+	}
+	if !ledgerstate.TransactionBalancesValid(consumedOutputs, tx.Essence().Outputs()) {
+		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: "sum of consumed and spent balances is not 0"})
+	}
+
+	// check unlock blocks validity
+	if !ledgerstate.UnlockBlocksValid(consumedOutputs, tx) {
+		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: "spending of referenced consumedOutputs is not authorized"})
+	}
+
+	// check if transaction is too old
+	if tx.Essence().Timestamp().Before(clock.SyncedTime().Add(-tangle.MaxReattachmentTimeMin)) {
+		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: fmt.Sprintf("transaction timestamp is older than MaxReattachmentTime (%s) and cannot be issued", tangle.MaxReattachmentTimeMin)})
+	}
+
+	// if transaction is in the future we wait until the time arrives
+	if tx.Essence().Timestamp().After(clock.SyncedTime()) {
+		time.Sleep(tx.Essence().Timestamp().Sub(clock.SyncedTime()) + 1*time.Nanosecond)
 	}
 
 	issueTransaction := func() (*tangle.Message, error) {