+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# v0.4.0 - 2021-02-26
+* Remove the value Tangle
+* Add approval switch
+* Add Markers integration
+* Add new message booker
+* Add new ledger state
+* Add new opinion former
+* Add new Tip manager
+* Add new flow unit tests
+* Add remote spammer tool
+* Add sendTransaction timestamp validity check
+* Add clock Since method
+* Add docs actions
+* Add Tangle width debug tool option
+* Update client-lib
+* Update wallet
+* Update web-API
+* Update Grafana dashboard to show MPS for entire data-flow
+* Update to latest hive.go
+* Refactor integration tests
+* Refactor transaction validity check
+* Refactor cli-wallet
+* Refactor sendTransactionByJson API
+* Refactor inclusion state
+* Refactor TransactionConfirmed event
+* Refactor transaction visualization in the local dashboard
+* Refactor scheduler
+* Refactor snapshot script
+* Fix clock time usage
+* Fix wrong handler in the MessageInvalidEvent
+* Decrease cache time
+* **Breaking**: bumps network and database versions
+
 # v0.3.6 - 2021-02-12
 * Finalize Payload layout
 * Update dRNG with finalized payload layout